refactor(cli): extract shared frontmatter parser, keep compat shims (Phase 2, p23-d)
